Moved the actor into the internal package
